*: separate ColumnsCommon
