+++ dbms/src/Columns/ColumnFixedString.cpp
@@ -13,7 +13,8 @@
 // limitations under the License.
 
 #include <Columns/ColumnFixedString.h>
-#include <Columns/ColumnsCommon.h>
+#include <Columns/countBytesInFilter.h>
+#include <Columns/filterColumn.h>
 #include <Common/Arena.h>
 #include <Common/HashTable/Hash.h>
 #include <Common/SipHash.h>
