+++ dbms/src/Columns/ColumnVector.cpp
@@ -13,7 +13,8 @@
 // limitations under the License.
 
 #include <Columns/ColumnVector.h>
-#include <Columns/ColumnsCommon.h>
+#include <Columns/countBytesInFilter.h>
+#include <Columns/filterColumn.h>
 #include <Common/Arena.h>
 #include <Common/Exception.h>
 #include <Common/HashTable/Hash.h>
