+++ dbms/src/Columns/filterColumn.cpp
@@ -13,209 +13,20 @@
 // limitations under the License.
 
 #include <Columns/ColumnDecimal.h>
-#include <Columns/ColumnsCommon.h>
-#include <Columns/IColumn.h>
+#include <Columns/ColumnUtil.h>
+#include <Columns/countBytesInFilter.h>
+#include <Columns/filterColumn.h>
 #include <common/memcpy.h>
-
-#if defined(__aarch64__) && defined(__ARM_NEON)
-#include <arm_neon.h>
-#endif
 
 #include <bit>
 
-#ifdef TIFLASH_ENABLE_AVX_SUPPORT
-ASSERT_USE_AVX2_COMPILE_FLAG
-#endif
-
-namespace DB
-{
-
-namespace
-{
-inline UInt64 ToBits64(const UInt8 * bytes64)
-{
-#if defined(__AVX512F__) && defined(__AVX512BW__)
-    const __m512i vbytes = _mm512_loadu_si512(reinterpret_cast<const void *>(bytes64));
-    UInt64 res = _mm512_testn_epi8_mask(vbytes, vbytes);
-#elif defined(__AVX2__)
-    const auto check_block = _mm256_setzero_si256();
-    uint64_t mask0 = mem_utils::details::get_block32_cmp_eq_mask(bytes64, check_block);
-    uint64_t mask1
-        = mem_utils::details::get_block32_cmp_eq_mask(bytes64 + mem_utils::details::BLOCK32_SIZE, check_block);
-    auto res = mask0 | (mask1 << mem_utils::details::BLOCK32_SIZE);
-#elif defined(__SSE2__)
-    const auto zero16 = _mm_setzero_si128();
-    UInt64 res = static_cast<UInt64>(_mm_movemask_epi8(
-                     _mm_cmpeq_epi8(_mm_loadu_si128(reinterpret_cast<const __m128i *>(bytes64)), zero16)))
-        | (static_cast<UInt64>(_mm_movemask_epi8(
-               _mm_cmpeq_epi8(_mm_loadu_si128(reinterpret_cast<const __m128i *>(bytes64 + 16)), zero16)))
-           << 16)
-        | (static_cast<UInt64>(_mm_movemask_epi8(
-               _mm_cmpeq_epi8(_mm_loadu_si128(reinterpret_cast<const __m128i *>(bytes64 + 32)), zero16)))
-           << 32)
-        | (static_cast<UInt64>(_mm_movemask_epi8(
-               _mm_cmpeq_epi8(_mm_loadu_si128(reinterpret_cast<const __m128i *>(bytes64 + 48)), zero16)))
-           << 48);
-#elif defined(__aarch64__) && defined(__ARM_NEON)
-    const uint8x16_t bitmask
-        = {0x01, 0x02, 0x4, 0x8, 0x10, 0x20, 0x40, 0x80, 0x01, 0x02, 0x4, 0x8, 0x10, 0x20, 0x40, 0x80};
-    const auto * src = reinterpret_cast<const unsigned char *>(bytes64);
-    const uint8x16_t p0 = vceqzq_u8(vld1q_u8(src));
-    const uint8x16_t p1 = vceqzq_u8(vld1q_u8(src + 16));
-    const uint8x16_t p2 = vceqzq_u8(vld1q_u8(src + 32));
-    const uint8x16_t p3 = vceqzq_u8(vld1q_u8(src + 48));
-    uint8x16_t t0 = vandq_u8(p0, bitmask);
-    uint8x16_t t1 = vandq_u8(p1, bitmask);
-    uint8x16_t t2 = vandq_u8(p2, bitmask);
-    uint8x16_t t3 = vandq_u8(p3, bitmask);
-    uint8x16_t sum0 = vpaddq_u8(t0, t1);
-    uint8x16_t sum1 = vpaddq_u8(t2, t3);
-    sum0 = vpaddq_u8(sum0, sum1);
-    sum0 = vpaddq_u8(sum0, sum0);
-    UInt64 res = vgetq_lane_u64(vreinterpretq_u64_u8(sum0), 0);
-#else
-    UInt64 res = 0;
-    for (size_t i = 0; i < 64; ++i)
-        res |= static_cast<UInt64>(0 == bytes64[i]) << i;
-#endif
-    return ~res;
-}
-
-constexpr size_t FILTER_SIMD_BYTES = 64;
-
-/// If mask is a number of this kind: [0]*[1]+ function returns the length of the cluster of 1s.
-/// Otherwise it returns the special value: 0xFF.
-/// Note: mask must be non-zero.
-inline UInt8 prefixToCopy(UInt64 mask)
-{
-    static constexpr UInt64 all_match = 0xFFFFFFFFFFFFFFFFULL;
-    if (mask == all_match)
-        return 64;
-    /// std::countl_zero count from the most significant bit of mask, corresponding to the tail of the original filter.
-    /// If only the tail of the original filter is zero, we can copy the prefix directly.
-    /// The length of tail zero if `leading_zeros`, so the length of the prefix to copy is 64 - #(leading zeroes).
-    const UInt64 leading_zeroes = std::countl_zero(mask);
-    if (mask == ((all_match << leading_zeroes) >> leading_zeroes))
-        return 64 - leading_zeroes;
-    else
-        return 0xFF;
-}
-
-inline UInt8 suffixToCopy(UInt64 mask)
-{
-    const auto prefix_to_copy = prefixToCopy(~mask);
-    return prefix_to_copy >= 64 ? prefix_to_copy : 64 - prefix_to_copy;
-}
-} // namespace
-
-
-ALWAYS_INLINE inline static size_t CountBytesInFilter(const UInt8 * filt, size_t start, size_t end)
-{
-#if defined(__AVX2__)
-    size_t size = end - start;
-    auto zero_cnt = mem_utils::details::avx2_byte_count(reinterpret_cast<const char *>(filt + start), size, 0);
-    return size - zero_cnt;
-#else
-    size_t count = 0;
-
-    /** NOTE: In theory, `filt` should only contain zeros and ones.
-      * But, just in case, here the condition > 0 (to signed bytes) is used.
-      * It would be better to use != 0, then this does not allow SSE2.
-      */
-
-    const char * pos = reinterpret_cast<const char *>(filt);
-    pos += start;
-
-    const char * end_pos = pos + (end - start);
-    for (; pos < end_pos; ++pos)
-        count += *pos != 0;
-
-    return count;
-#endif
-}
-
-size_t countBytesInFilter(const UInt8 * filt, size_t sz)
-{
-    return CountBytesInFilter(filt, 0, sz);
-}
-
-size_t countBytesInFilter(const UInt8 * filt, size_t start, size_t sz)
-{
-    return CountBytesInFilter(filt, start, start + sz);
-}
-
-size_t countBytesInFilter(const IColumn::Filter & filt, size_t start, size_t sz)
-{
-    return CountBytesInFilter(filt.data(), start, start + sz);
-}
-
-size_t countBytesInFilter(const IColumn::Filter & filt)
-{
-    return CountBytesInFilter(filt.data(), 0, filt.size());
-}
-
-static inline size_t CountBytesInFilterWithNull(const UInt8 * p1, const UInt8 * p2, size_t size)
-{
-    size_t count = 0;
-    for (size_t i = 0; i < size; ++i)
-    {
-        count += (p1[i] & ~p2[i]) != 0;
-    }
-    return count;
-}
-
-static inline size_t CountBytesInFilterWithNull(
-    const IColumn::Filter & filt,
-    const UInt8 * null_map,
-    size_t start,
-    size_t end)
-{
-    size_t count = 0;
-
-    /** NOTE: In theory, `filt` should only contain zeros and ones.
-      * But, just in case, here the condition > 0 (to signed bytes) is used.
-      * It would be better to use != 0, then this does not allow SSE2.
-      */
-
-    const auto * p1 = filt.data() + start;
-    const auto * p2 = null_map + start;
-    size_t size = end - start;
-
-#if defined(__SSE2__) || defined(__AVX2__)
-    for (; size >= 64;)
-    {
-        count += std::popcount(ToBits64(p1) & ~ToBits64(p2));
-        p1 += 64, p2 += 64;
-        size -= 64;
-    }
-#endif
-    count += CountBytesInFilterWithNull(p1, p2, size);
-    return count;
-}
-
-size_t countBytesInFilterWithNull(const IColumn::Filter & filt, const UInt8 * null_map)
-{
-    return CountBytesInFilterWithNull(filt, null_map, 0, filt.size());
-}
-
-size_t countBytesInFilterWithNull(const IColumn::Filter & filt, const UInt8 * null_map, size_t start, size_t sz)
-{
-    return CountBytesInFilterWithNull(filt, null_map, start, start + sz);
-}
-
-std::vector<size_t> countColumnsSizeInSelector(IColumn::ColumnIndex num_columns, const IColumn::Selector & selector)
-{
-    std::vector<size_t> counts(num_columns);
-    for (auto idx : selector)
-        ++counts[idx];
-
-    return counts;
-}
-
-namespace ErrorCodes
+namespace DB::ErrorCodes
 {
 extern const int SIZES_OF_COLUMNS_DOESNT_MATCH;
 }
+
+namespace DB
+{
 
 namespace
 {
