+++ dbms/src/Columns/filterColumn.h
@@ -16,22 +16,10 @@
 
 #include <Columns/IColumn.h>
 
-
 /// Common helper methods for implementation of different columns.
 
 namespace DB
 {
-/// Counts how many bytes of `filt` are greater than zero.
-size_t countBytesInFilter(const UInt8 * filt, size_t sz);
-size_t countBytesInFilter(const UInt8 * filt, size_t start, size_t sz);
-size_t countBytesInFilter(const IColumn::Filter & filt);
-size_t countBytesInFilter(const IColumn::Filter & filt, size_t start, size_t sz);
-size_t countBytesInFilterWithNull(const IColumn::Filter & filt, const UInt8 * null_map);
-size_t countBytesInFilterWithNull(const IColumn::Filter & filt, const UInt8 * null_map, size_t start, size_t sz);
-
-/// Returns vector with num_columns elements. vector[i] is the count of i values in selector.
-/// Selector must contain values from 0 to num_columns - 1. NOTE: this is not checked.
-std::vector<size_t> countColumnsSizeInSelector(IColumn::ColumnIndex num_columns, const IColumn::Selector & selector);
 
 /// The general implementation of `filter` function for ColumnArray and ColumnString.
 template <typename T>
