+++ dbms/src/Columns/tests/bench_column_filter.cpp
@@ -14,7 +14,7 @@
 
 
 #include <Columns/ColumnVector.h>
-#include <Columns/ColumnsCommon.h>
+#include <Columns/filterColumn.h>
 #include <Common/typeid_cast.h>
 #include <benchmark/benchmark.h>
